Standardizes DAG names, updates tags, and moves experiments and scratchwork to a dev_experiments directory.
