--- airflow/dags/cook_county/extract_and_load_parcel_locations.py
@@ -1,52 +1,0 @@
-import datetime as dt
-import logging
-
-from airflow.models.baseoperator import chain
-from airflow.decorators import dag
-from airflow.operators.empty import EmptyOperator
-from airflow.utils.trigger_rule import TriggerRule
-
-from cc_utils.socrata import SocrataTable
-from tasks.socrata_tasks import (
-    download_fresh_data,
-    fresher_source_data_available,
-    check_table_metadata,
-    load_data_tg,
-)
-
-
-task_logger = logging.getLogger("airflow.task")
-
-SOCRATA_TABLE = SocrataTable(table_id="c49d-89sn", table_name="cook_county_parcel_locations")
-
-
-@dag(
-    schedule="0 7 4 * *",
-    start_date=dt.datetime(2022, 11, 1),
-    catchup=False,
-    tags=["cook_county", "parcels", "dimension_table"],
-)
-def update_cc_parcel_locations_table():
-    POSTGRES_CONN_ID = "dwh_db_conn"
-
-    end_1 = EmptyOperator(task_id="end", trigger_rule=TriggerRule.NONE_FAILED)
-
-    metadata_1 = check_table_metadata(
-        socrata_table=SOCRATA_TABLE, conn_id=POSTGRES_CONN_ID, task_logger=task_logger
-    )
-    fresh_source_data_available_1 = fresher_source_data_available(
-        socrata_metadata=metadata_1, conn_id=POSTGRES_CONN_ID, task_logger=task_logger
-    )
-    extract_data_1 = download_fresh_data(task_logger=task_logger)
-    load_data_tg_1 = load_data_tg(
-        socrata_metadata=extract_data_1,
-        socrata_table=SOCRATA_TABLE,
-        conn_id=POSTGRES_CONN_ID,
-        task_logger=task_logger,
-    )
-
-    chain(metadata_1, fresh_source_data_available_1, extract_data_1, load_data_tg_1)
-    chain(metadata_1, fresh_source_data_available_1, end_1)
-
-
-update_cc_parcel_locations_table()
