+++ airflow/dags/cook_county/update_raw_cook_county_parcel_sales.py
@@ -1,39 +1,51 @@
 import datetime as dt
 import logging
 
+from airflow.models.baseoperator import chain
 from airflow.decorators import dag
-from airflow.operators.bash import BashOperator
 from airflow.operators.empty import EmptyOperator
 from airflow.utils.trigger_rule import TriggerRule
 
 from cc_utils.socrata import SocrataTable
-
+from tasks.socrata_tasks import (
+    download_fresh_data,
+    fresher_source_data_available,
+    check_table_metadata,
+    load_data_tg,
+)
 
 task_logger = logging.getLogger("airflow.task")
 
-# SOCRATA_TABLE = SocrataTable(table_id="wvhk-k5uv", table_name="cook_county_parcel_sales")
+SOCRATA_TABLE = SocrataTable(table_id="wvhk-k5uv", table_name="cook_county_parcel_sales")
 
 
 @dag(
-    schedule=None,
+    schedule="0 6 4 * *",
     start_date=dt.datetime(2022, 11, 1),
     catchup=False,
-    tags=["cook_county", "boundary_lines", "dimension_table", "geospatial"],
+    tags=["cook_county", "parcels", "fact_table", "data_raw"],
 )
 def update_data_raw_cook_county_parcel_sales():
-    # dbt_stage_table_1 = BashOperator(
-    #     task_id="dbt_stage_table",
-    #     bash_command="cd /opt/airflow/dbt && dbt run --select models/staging/stg_cc_parcel_sales.sql",
-    # )
-    dbt_update_data_raw_table_1 = BashOperator(
-        task_id="dbt_update_data_raw_table",
-        bash_command="""cd /opt/airflow/dbt && \
-            dbt run --select models/staging/cook_county_parcel_sales.sql""",
-    )
+    POSTGRES_CONN_ID = "dwh_db_conn"
+
     end_1 = EmptyOperator(task_id="end", trigger_rule=TriggerRule.NONE_FAILED)
 
-    # dbt_stage_table_1 >>
-    dbt_update_data_raw_table_1 >> end_1
+    metadata_1 = check_table_metadata(
+        socrata_table=SOCRATA_TABLE, conn_id=POSTGRES_CONN_ID, task_logger=task_logger
+    )
+    fresh_source_data_available_1 = fresher_source_data_available(
+        socrata_metadata=metadata_1, conn_id=POSTGRES_CONN_ID, task_logger=task_logger
+    )
+    extract_data_1 = download_fresh_data(task_logger=task_logger)
+    load_data_tg_1 = load_data_tg(
+        socrata_metadata=extract_data_1,
+        socrata_table=SOCRATA_TABLE,
+        conn_id=POSTGRES_CONN_ID,
+        task_logger=task_logger,
+    )
+
+    chain(metadata_1, fresh_source_data_available_1, extract_data_1, load_data_tg_1)
+    chain(metadata_1, fresh_source_data_available_1, end_1)
 
 
 update_data_raw_cook_county_parcel_sales()
